fix: metodo excluir projeto

+++ src/main/java/br/com/ifpr/edu/sdpe_backend/service/ProjetoService.java
@@ -1,11 +1,12 @@
 package br.com.ifpr.edu.sdpe_backend.service;
 
 import br.com.ifpr.edu.sdpe_backend.domain.*;
-//import br.com.ifpr.edu.sdpe_backend.domain.DTO.EstatisticaDTO;
+import br.com.ifpr.edu.sdpe_backend.repository.ContatoRepository;
 import br.com.ifpr.edu.sdpe_backend.repository.InstituicaoEnsinoRepository;
 import br.com.ifpr.edu.sdpe_backend.repository.ProjetoRepository;
 import br.com.ifpr.edu.sdpe_backend.repository.VisualizacaoRepository;
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -35,6 +36,8 @@ public class ProjetoService {
     private final CoordenadorService coordenadorService;
 
     private final VisualizacaoRepository visualizacaoRepository;
+
+    private final ContatoRepository contatoRepository;
 
     private final Path rootLocation = Paths.get("uploads");
 
@@ -98,7 +101,6 @@ public class ProjetoService {
 
 
                 if (busca.isPresent()) {
-                    // Se já existe, usamos a do banco (ignorando a descrição nova para não sobrescrever dados antigos sem querer)
                     instituicaoFinal = busca.get();
                 } else {
                     instituicaoFinal = instituicaoEnsinoRepository.save(input);
@@ -165,26 +167,34 @@ public class ProjetoService {
         return existente;
     }
 
+    @Transactional
     public void excluir(Long id) {
-        Projeto projeto = this.projetoRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException("Projeto a ser atualizado não encontrado"));
+        Projeto projeto = this.buscarPorId(id);
 
-        if (projeto.getImagemPath() != null && !projeto.getImagemPath().trim().isEmpty())
-            deletarArquivoFisico(projeto.getImagemPath());
+        visualizacaoRepository.deleteByProjetoId(id);
 
-        if (projeto.getDocumentoPath() != null && !projeto.getDocumentoPath().trim().isEmpty())
-            deletarArquivoFisico(projeto.getDocumentoPath());
+        List<Contato> contatos = contatoRepository.findByProjetoId(id);
+        if (!contatos.isEmpty()) contatoRepository.deleteAll(contatos);
+
+        deletarArquivoFisico(projeto.getImagemPath());
+        deletarArquivoFisico(projeto.getDocumentoPath());
 
         this.projetoRepository.deleteById(id);
     }
 
-    private void deletarArquivoFisico(String nomeArquivo) {
-        try {
-            Path arquivo = rootLocation.resolve(nomeArquivo);
+    private void deletarArquivoFisico(String caminhoOuUrl) {
+        if (caminhoOuUrl != null && !caminhoOuUrl.isBlank()) {
+            try {
+                String nomeArquivo = caminhoOuUrl;
+                if (caminhoOuUrl.contains("/")) {
+                    nomeArquivo = caminhoOuUrl.substring(caminhoOuUrl.lastIndexOf("/") + 1);
+                }
 
-            Files.deleteIfExists(arquivo);
-        } catch (IOException e) {
-            System.err.println("Erro ao deletar arquivo físico (" + nomeArquivo + "): " + e.getMessage());
+                Path arquivo = rootLocation.resolve(nomeArquivo);
+                Files.deleteIfExists(arquivo);
+            } catch (Exception e) {
+                System.err.println("Erro ao deletar arquivo físico (" + caminhoOuUrl + "): " + e.getMessage());
+            }
         }
     }
 
@@ -264,7 +274,7 @@ public class ProjetoService {
         Projeto projeto = buscarPorId(idProjeto);
 
         contato.setProjeto(projeto);
-        
+
         return contatoService.salvar(contato);
     }
